// frontend/src/core/api/generated/model/createTenantRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Constructor ERP API
 * OpenAPI spec version: 1.0
 */
import type { PaymentRequest } from './paymentRequest'

export interface CreateTenantRequest {
  /**
   * @minLength 0
   * @maxLength 120
   */
  name: string
  /**
   * @minLength 0
   * @maxLength 20
   * @nullable
   */
  phone?: string | null
  /**
   * @minLength 0
   * @maxLength 254
   * @nullable
   */
  email?: string | null
  /**
   * @minLength 0
   * @maxLength 60
   * @nullable
   */
  city?: string | null
  planId: string
  /**
   * @minimum 1
   * @maximum 36
   */
  months: number
  /** @nullable */
  startsOn?: string | null
  payment?: PaymentRequest
  /** @nullable */
  salesRequestId?: string | null
}