// frontend/src/core/api/generated/model/planAdminView.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Constructor ERP API
 * OpenAPI spec version: 1.0
 */

export interface PlanAdminView {
  id: string
  code: string
  name: string
  /** @nullable */
  tagline?: string | null
  /** @nullable */
  monthlyPrice?: number | null
  currency: string
  /** @nullable */
  maxUsers?: number | null
  /** @nullable */
  maxSites?: number | null
  highlighted: boolean
  visible: boolean
  status: string
  features: string[]
  activeTenants: number
}